+++ src/screens/USDCoinScreen/ListItem/ListItem.types.ts
@@ -1,0 +1,8 @@
+import { GestureResponderEvent } from 'react-native';
+
+export interface ListItemProps {
+	onPress: (event: GestureResponderEvent) => void;
+	label: string;
+	selected: boolean;
+	token?: string;
+}
